+++ packages/icons/docs/stories/helpers/iconsArray.js
@@ -80,6 +80,9 @@ import {
     InteractiveExamples,
     YoutubeIcon,
     MediumIcon,
+    Filter,
+    PlaneFilled,
+    Update,
 } from '../../../src';
 
 export default [
@@ -488,5 +491,20 @@ export default [
         name: 'mediumIcon',
         category: 'custom',
         icon: MediumIcon,
+    },
+    {
+        name: 'filter',
+        category: 'standard',
+        icon: Filter,
+    },
+    {
+        name: 'planeFilled',
+        category: 'standard',
+        icon: PlaneFilled,
+    },
+    {
+        name: 'update',
+        category: 'standard',
+        icon: Update,
     },
 ];