+++ packages/icons/src/index.d.ts
@@ -76,3 +76,5 @@ export { default as OneFilled } from './components/oneFilled';
 export { default as List } from './components/list';
 export { default as Seat } from './components/seat';
 export { default as FrontCar } from './components/frontCar';
+export { default as Upadte } from './components/update';
+export { default as PlaneFilled } from './components/planeFilled';